feature: do not delete user input in case sending fails
In case the sending callback function returns "false",
then sending the message is regarded as failed. In such cases
the user input field is not deleted. Then the user might try
to resolve sending issues and try again, without loosing
its message.

Previously the input field was deleted in any case.
If the message is sent to a remote server via network,
the network might be down currently and even on
purpose. Trying to send a message in that situation
had lead to losing the message.

Now, the message is retained until the callback to send
the message returns. The return value is evaluated
for any sign of success. On success, the input field is
cleared and in failure, its content is maintained.